+++ src/application/cargue-masivo/cargue-masivo.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CargueMasivoController } from './cargue-masivo.controller';
 import { CargueMasivoService } from './cargue-masivo.service';
-import { NuxeoService } from 'src/utils/axios/nuxeo.service';
+import { NuxeoService } from 'src/shared/utils/nuxeo/nuxeo.service';
 
 describe('CargueMasivoController', () => {
   let controller: CargueMasivoController;
