+++ src/application/cargue-masivo/cargue-masivo.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { CargueMasivoController } from './cargue-masivo.controller';
 import { CargueMasivoService } from './cargue-masivo.service';
 import { HttpModule } from '@nestjs/axios';
-import { NuxeoModule } from 'src/utils/axios/nuxeo.module';
+import { NuxeoModule } from 'src/shared/utils/nuxeo/nuxeo.module';
 
 @Module({
   imports: [HttpModule, NuxeoModule],
