+++ src/shared/utils/base64.utils.ts
