+++ src/shared/utils/nuxeo/nuxeo.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
-import { NuxeoService } from "./nuxeo.service";
-import { GestorDocumentalService } from "./gestor-documental.service";
+import { NuxeoService } from "src/shared/utils/nuxeo/nuxeo.service";
+import { GestorDocumentalService } from "../../services/gestor-documental/gestor-documental.service";
 
 @Module({
   providers: [NuxeoService, GestorDocumentalService],
